% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_density.R
\name{calc_density}
\alias{calc_density}
\title{Create density curves}
\usage{
calc_density(new_data_with_pred, var_name, vec_y_expand = NULL)
}
\arguments{
\item{new_data_with_pred}{Data with prediction to make density calculations on}

\item{var_name}{String specifying which variable to calculate density}

\item{vec_y_expand}{Optional values to expand y-axis}
}
\value{
Density data for plotting
}
\description{
Create density curves
}
\details{
It should be amenable to any 1st-order model without interaction terms,
however not implemented yet, such as handling \code{predict()} function output
for binary endpoint
}
\keyword{internal}
