% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_density.R
\name{calc_ice_quantile}
\alias{calc_ice_quantile}
\title{Create ICE curves at quantiles}
\usage{
calc_ice_quantile(
  object,
  new_data,
  var_name,
  qtiles = seq(0, 1, by = 0.1),
  task = "regression"
)
}
\arguments{
\item{object}{Object of class sculpture (rough, detailed)}

\item{new_data}{Data to make quantiles on}

\item{var_name}{String specifying which variable to generate ICE}

\item{qtiles}{Quantiles to generate ICE curves}

\item{task}{Prediction task type (regression or classification)}
}
\value{
Predictions
}
\description{
Create ICE curves at quantiles
}
\details{
It should be amenable to any 1st-order model without interaction terms,
however not implemented yet, such as handling \code{predict()} function output
for binary endpoint
}
\keyword{internal}
