% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_density.R
\name{expand_range}
\alias{expand_range}
\title{Expand the range of values for density plot}
\usage{
expand_range(
  x,
  expand_left_side = 0.1,
  expand_right_side = 0.2,
  type = c("relative", "absolute")
)
}
\arguments{
\item{x}{numeric vector}

\item{expand_left_side}{Fraction to expand on left hand side}

\item{expand_right_side}{Fraction to expand on right hand side}
}
\value{
Vector of 2 values
}
\description{
Expand the range of values for density plot
}
\keyword{internal}
