% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{g_ice}
\alias{g_ice}
\title{Plot centered ICE profiles with centered PDP curves}
\usage{
g_ice(
  object,
  centered = TRUE,
  show_PDP = TRUE,
  coloured = FALSE,
  rug_sides = "b",
  missings_spec = missings_specification(),
  facet_spec = facet_specification(),
  logodds_to_prob = FALSE
)
}
\arguments{
\item{object}{Object of classes \code{rough} and \code{sculpture}.}

\item{centered}{\code{logical}, centered ice plots? Defaults to \code{TRUE}.}

\item{show_PDP}{\code{logical}, show PDP line? Defaults to \code{TRUE}.}

\item{coloured}{\code{logical}, coloured curves? Defaults to \code{FALSE}.}

\item{rug_sides}{"" for none, "b", for bottom, "trbl" for all 4 sides (see \code{geom_rug})}

\item{missings_spec}{Object of class \code{missings_specificatoin}.}

\item{facet_spec}{Object of class \code{facet_specificatoin}.}

\item{logodds_to_prob}{(\code{logical}) Only valid for binary response and sculptures built on
the log-odds scale. Defaults to \code{FALSE} (i.e. no effect).
If \code{TRUE}, then the y-values are transformed through inverse logit function 1 / (1 + exp(-x)).}
}
\value{
List of \code{ggplot}s (one for continuous features, one for discrete).
}
\description{
Plot centered ICE profiles with centered PDP curves
}
\examples{
df <- mtcars
df$vs <- as.factor(df$vs)
model <- rpart::rpart(
  hp ~ mpg + carb + vs,
  data = df,
  control = rpart::rpart.control(minsplit = 10)
)
model_predict <- function(x) predict(model, newdata = x)
covariates <- c("mpg", "carb", "vs")
pm <- sample_marginals(df[covariates], n = 50, seed = 5)

rs <- sculpt_rough(
  dat = pm,
  model_predict_fun = model_predict,
  n_ice = 10,
  seed = 1,
  verbose = 0
)

g_ice(rs)$continuous

}
