% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{missings_specification}
\alias{missings_specification}
\title{Instructions for missings vizualisations}
\usage{
missings_specification(
  vline = FALSE,
  hline = FALSE,
  values = NULL,
  drop_from_plot = FALSE
)
}
\arguments{
\item{vline}{(\code{logical}) Should the vertical line be shown? Defaults to \code{FALSE}.}

\item{hline}{(\code{logical}) Should the horizontal line be shown? Defaults to \code{FALSE}.}

\item{values}{(\code{NULL}) or single value or a named vector.
Specifies the value(-s) that stand for the missing values.
If \code{NULL}, then no missing value handling is carried out.
If single value, then it is assumed that this value is used for flagging missing values across
all continuous variables.
If named vector, then the names are used to refer to continuous variables and the values for
flagging missing values in that variable.}

\item{drop_from_plot}{(\code{logical}) Should the missing values be dropped from plot?
Defaults to \code{FALSE}.}
}
\value{
List of class \code{missings_specification}.
}
\description{
Instructions for missings vizualisations
}
\examples{
\dontrun{
g_ice(
  sculpture,
  missings_spec = missings_specification(
    vline = TRUE, # show vertical line
    values = -1 # NAs in all continuous variables displayed as -1
  )
)
}

}
