% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parallel_set}
\alias{parallel_set}
\alias{parallel_end}
\title{Set and end parallel computation}
\usage{
parallel_set(num_cores = 10, cluster_type = "fork")

parallel_end()
}
\arguments{
\item{num_cores}{(\code{integer}) Number of cores.}

\item{cluster_type}{(\code{character}) Type of cluster. One of \code{c("fork", "psock")}.}
}
\description{
Set and end parallel computation
}
\examples{
\dontrun{
parallel_set(num_cores = 2)
# now the code will run on parallel with 2 cores
parallel_end()
# now the code will run sequentially
}
}
