% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sculpting.R
\name{sample_marginals}
\alias{sample_marginals}
\title{Sample product marginals dataset}
\usage{
sample_marginals(dat, n, seed = NULL)
}
\arguments{
\item{dat}{Data.frame to sample from, must include only covariates.}

\item{n}{Number of observations to sample.}

\item{seed}{\code{NULL} or seed for exact reproducibility.}
}
\value{
\code{data.frame} with same number of columns and \code{n} rows.
}
\description{
Sample product marginals dataset
}
\details{
The product marginals dataset is a grid of values that is sampled independently
per each column (feature) from the original dataset.
The aim here is to disentangle the correlations between features and assess
how each feature affects the model predictions individually.
It will not contain new values per column, but it may contain new combinations of values not
seen in the original data.
One can also check how the model behaves if there are unseen observations
(new combination of features).
Note that the use of the product marginal dataset for model sculpting only works
if the features are approximately additive for model predictions.
In the quite rare case when they are not, the sculpted models using the product marginal
dataset is expected to have significantly lower performance and
the conclusions may be misleading.

One can also try using the original data instead of the product marginals for model
sculpting and see how the results differ.
}
\examples{
sample_marginals(mtcars, n = 5, seed = 543)
}
