% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sculpting.R
\name{sculpt_polished}
\alias{sculpt_polished}
\title{Create a polished model}
\usage{
sculpt_polished(object, k = NULL, vars = NULL)
}
\arguments{
\item{object}{Object of class \code{sculpture}, either \code{rough} or \code{detailed}.}

\item{k}{Number of most important variables to keep.}

\item{vars}{Vector of variables to keep.}
}
\value{
Object of classes \code{rough} / \code{detailed} and \code{sculpture}.
}
\description{
Create a polished model
}
\examples{
df <- mtcars
df$vs <- as.factor(df$vs)
model <- rpart::rpart(
  hp ~ mpg + carb + vs,
  data = df,
  control = rpart::rpart.control(minsplit = 10)
)
model_predict <- function(x) predict(model, newdata = x)
covariates <- c("mpg", "carb", "vs")
pm <- sample_marginals(df[covariates], n = 50, seed = 5)

rs <- sculpt_rough(
  dat = pm,
  model_predict_fun = model_predict,
  n_ice = 10,
  seed = 1,
  verbose = 0
)

ds <- sculpt_detailed_gam(rs)

# this keeps only "mpg"
ps <- sculpt_polished(ds, k = 1)

}
